// package/code-push-lib.d.ts
interface IPackage {
    deploymentKey: string;
    description: string;
    label: string;
    appVersion: string;
    isMandatory: boolean;
    packageHash: string;
    packageSize: number;
    failedInstall: boolean;
}
// ...
interface ILocalPackage extends IPackage {
    /**
     * The local storage path where this package is located.
     */
    localPath: string;

    /**
     * Indicates if the current application run is the first one after the package was
     * applied.
     */
    isFirstRun: boolean;

    /**
     * Applies this package to the application. The application will be reloaded with
     * this package and on every application launch this package will be loaded.
     * On the first run after the update, the application will wait for a
     * codePush.notifyApplicationReady() call. Once this call is made, the install
     * operation is considered a success.
     * Otherwise, the install operation will be marked as failed, and the application is
     * reverted to its previous version on the next run.
     *
     * @param installSuccess Callback invoked if the install operation succeeded.
     * @param installError Optional callback inovoked in case of an error.
     * @param installOptions Optional parameter used for customizing the installation
     * behavior.
     */
    install(installSuccess: SuccessCallback<InstallMode>,
            installError?: ErrorCallback,
            installOptions?: InstallOptions): void;

    installWithPromise(installProgressEvent: EventEmitter<null | number>,
            installOptions?: InstallOptions): Promise<InstallMode>;
}
// ...
interface SuccessCallback<T> { (result?: T): void;
}
interface ErrorCallback { (error?: Error): void;
}
// ...
declare enum InstallMode {
    /**
     * The update will be applied to the running application immediately.
     * The application will be reloaded with the new content immediately.
     */
    IMMEDIATE,

        /**
         * The update is downloaded but not installed immediately.
         * The new content will be available the next time the application is started.
         */
    ON_NEXT_RESTART,

        /**
         * The udpate is downloaded but not installed immediately.
         * The new content will be available the next time the application is
         * resumed or restarted, whichever event happends first.
         */
    ON_NEXT_RESUME
}

/**
 * Defines the install operation options.
 */
interface InstallOptions {
    /**
     * Used to specify the InstallMode used for the install operation.
     * This is optional and defaults to InstallMode.ON_NEXT_RESTART.
     */
    installMode?: InstallMode;

    /**
     * If installMode === ON_NEXT_RESUME, the minimum amount of time (in seconds)
     * which needs to pass with the app in the background before an update install
     * occurs when the app is resumed.
     */
    minimumBackgroundDuration?: number;

    /**
     * Used to specify the InstallMode used for the install operation if the update
     * is mandatory. This is optional and defaults to InstallMode.IMMEDIATE.
     */
    mandatoryInstallMode?: InstallMode;
}